requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# Image processing
opencv-python>=4.8.0
scikit-image>=0.21.0

# Metrics and evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars and utilities
tqdm>=4.66.0
pandas>=2.0.0

# Diffusion models
diffusers>=0.21.4
transformers>=4.30.0
accelerate>=0.20.0

# Additional dependencies
einops>=0.6.1
timm>=0.9.0

# Optional but recommended
tensorboard>=2.13.0